+++ src/main/java/net/geforcemods/securitycraft/blockentities/LaserBlockBlockEntity.java
@@ -7,22 +7,17 @@ import java.util.Map;
 import java.util.Map.Entry;
 
 import it.unimi.dsi.fastutil.objects.Object2BooleanArrayMap;
-import net.geforcemods.securitycraft.ClientHandler;
 import net.geforcemods.securitycraft.ConfigHandler;
 import net.geforcemods.securitycraft.SCContent;
-import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.api.ILinkedAction;
 import net.geforcemods.securitycraft.api.LinkableBlockEntity;
 import net.geforcemods.securitycraft.api.Option;
 import net.geforcemods.securitycraft.api.Option.DisabledOption;
 import net.geforcemods.securitycraft.api.Option.IgnoreOwnerOption;
 import net.geforcemods.securitycraft.api.Owner;
-import net.geforcemods.securitycraft.blocks.DisguisableBlock;
 import net.geforcemods.securitycraft.blocks.LaserBlock;
 import net.geforcemods.securitycraft.items.ModuleItem;
 import net.geforcemods.securitycraft.misc.ModuleType;
-import net.geforcemods.securitycraft.models.DisguisableDynamicBakedModel;
-import net.geforcemods.securitycraft.network.client.RefreshDisguisableModel;
 import net.geforcemods.securitycraft.util.BlockUtils;
 import net.minecraft.Util;
 import net.minecraft.core.BlockPos;
@@ -31,13 +26,8 @@ import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.Tag;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.block.state.properties.BlockStateProperties;
-import net.minecraft.world.level.material.Fluids;
 import net.minecraftforge.client.model.data.IModelData;
-import net.minecraftforge.client.model.data.ModelDataMap;
-import net.minecraftforge.network.PacketDistributor;
 
 public class LaserBlockBlockEntity extends LinkableBlockEntity {
 	private DisabledOption disabled = new DisabledOption(false) {
@@ -106,27 +96,10 @@ public class LaserBlockBlockEntity extends LinkableBlockEntity {
 			else if (option.getName().equals("ignoreOwner"))
 				ignoreOwner.copy(option);
 		}
-		else if (action instanceof ILinkedAction.ModuleInserted moduleInserted) {
-			ItemStack module = moduleInserted.stack();
-			boolean toggled = moduleInserted.wasModuleToggled();
-
-			insertModule(module, toggled);
-
-			if (moduleInserted.module().getModuleType() == ModuleType.DISGUISE)
-				onInsertDisguiseModule(module, toggled);
-		}
-		else if (action instanceof ILinkedAction.ModuleRemoved moduleRemoved) {
-			ModuleType module = moduleRemoved.moduleType();
-			ItemStack moduleStack = getModule(module);
-			boolean toggled = moduleRemoved.wasModuleToggled();
-
-			removeModule(module, toggled);
-
-			if (module == ModuleType.DISGUISE)
-				onRemoveDisguiseModule(moduleStack, toggled);
-			else if (module == ModuleType.REDSTONE)
-				onRemoveRedstoneModule();
-		}
+		else if (action instanceof ILinkedAction.ModuleInserted moduleInserted)
+			insertModule(moduleInserted.stack(), moduleInserted.wasModuleToggled());
+		else if (action instanceof ILinkedAction.ModuleRemoved moduleRemoved)
+			removeModule(moduleRemoved.moduleType(), moduleRemoved.wasModuleToggled());
 		else if (action instanceof ILinkedAction.OwnerChanged ownerChanged) {
 			Owner owner = ownerChanged.newOwner();
 
@@ -151,7 +124,7 @@ public class LaserBlockBlockEntity extends LinkableBlockEntity {
 		super.onModuleInserted(stack, module, toggled);
 
 		if (module == ModuleType.DISGUISE)
-			onInsertDisguiseModule(stack, toggled);
+			DisguisableBlockEntity.onDisguiseModuleInserted(this, stack, toggled);
 		else if (module == ModuleType.SMART)
 			applyExistingSideConfig();
 	}
@@ -161,71 +134,21 @@ public class LaserBlockBlockEntity extends LinkableBlockEntity {
 		super.onModuleRemoved(stack, module, toggled);
 
 		if (module == ModuleType.DISGUISE)
-			onRemoveDisguiseModule(stack, toggled);
-		else if (module == ModuleType.REDSTONE)
-			onRemoveRedstoneModule();
+			DisguisableBlockEntity.onDisguiseModuleRemoved(this, stack, toggled);
+		else if (module == ModuleType.REDSTONE) {
+			if (getBlockState().getValue(LaserBlock.POWERED)) {
+				level.setBlockAndUpdate(worldPosition, getBlockState().setValue(LaserBlock.POWERED, false));
+				BlockUtils.updateIndirectNeighbors(level, worldPosition, SCContent.LASER_BLOCK.get());
+			}
+		}
 		else if (module == ModuleType.SMART)
 			applyExistingSideConfig();
-	}
-
-	private void onInsertDisguiseModule(ItemStack stack, boolean toggled) {
-		BlockState state = getBlockState();
-
-		if (!level.isClientSide) {
-			SecurityCraft.channel.send(PacketDistributor.TRACKING_CHUNK.with(() -> level.getChunkAt(worldPosition)), new RefreshDisguisableModel(worldPosition, true, stack, toggled));
-
-			if (state.hasProperty(BlockStateProperties.WATERLOGGED) && state.getValue(BlockStateProperties.WATERLOGGED)) {
-				level.scheduleTick(worldPosition, Fluids.WATER, Fluids.WATER.getTickDelay(level));
-				level.updateNeighborsAt(worldPosition, state.getBlock());
-			}
-		}
-		else {
-			ClientHandler.putDisguisedBeRenderer(this, stack);
-
-			if (state.getLightEmission(level, worldPosition) > 0)
-				level.getChunkSource().getLightEngine().checkBlock(worldPosition);
-		}
-	}
-
-	private void onRemoveDisguiseModule(ItemStack stack, boolean toggled) {
-		if (!level.isClientSide) {
-			BlockState state = getBlockState();
-
-			SecurityCraft.channel.send(PacketDistributor.TRACKING_CHUNK.with(() -> level.getChunkAt(worldPosition)), new RefreshDisguisableModel(worldPosition, false, stack, toggled));
-
-			if (state.hasProperty(BlockStateProperties.WATERLOGGED) && state.getValue(BlockStateProperties.WATERLOGGED)) {
-				level.scheduleTick(worldPosition, Fluids.WATER, Fluids.WATER.getTickDelay(level));
-				level.updateNeighborsAt(worldPosition, state.getBlock());
-			}
-		}
-		else {
-			ClientHandler.DISGUISED_BLOCK_RENDER_DELEGATE.removeDelegateOf(this);
-			DisguisableBlock.getDisguisedBlockStateFromStack(stack).ifPresent(disguisedState -> {
-				if (disguisedState.getLightEmission(level, worldPosition) > 0)
-					level.getChunkSource().getLightEngine().checkBlock(worldPosition);
-			});
-		}
-	}
-
-	private void onRemoveRedstoneModule() {
-		if (getBlockState().getValue(LaserBlock.POWERED)) {
-			level.setBlockAndUpdate(worldPosition, getBlockState().setValue(LaserBlock.POWERED, false));
-			BlockUtils.updateIndirectNeighbors(level, worldPosition, SCContent.LASER_BLOCK.get());
-		}
 	}
 
 	@Override
 	public void handleUpdateTag(CompoundTag tag) {
 		super.handleUpdateTag(tag);
-
-		if (level != null && level.isClientSide) {
-			ItemStack stack = getModule(ModuleType.DISGUISE);
-
-			if (!stack.isEmpty())
-				ClientHandler.putDisguisedBeRenderer(this, stack);
-			else
-				ClientHandler.DISGUISED_BLOCK_RENDER_DELEGATE.removeDelegateOf(this);
-		}
+		DisguisableBlockEntity.onHandleUpdateTag(this);
 	}
 
 	@Override
@@ -241,9 +164,7 @@ public class LaserBlockBlockEntity extends LinkableBlockEntity {
 	@Override
 	public void setRemoved() {
 		super.setRemoved();
-
-		if (level.isClientSide)
-			ClientHandler.DISGUISED_BLOCK_RENDER_DELEGATE.removeDelegateOf(this);
+		DisguisableBlockEntity.onSetRemoved(this);
 	}
 
 	@Override
@@ -262,7 +183,7 @@ public class LaserBlockBlockEntity extends LinkableBlockEntity {
 
 	@Override
 	public IModelData getModelData() {
-		return new ModelDataMap.Builder().withInitial(DisguisableDynamicBakedModel.DISGUISED_STATE, Blocks.AIR.defaultBlockState()).build();
+		return DisguisableBlockEntity.DEFAULT_MODEL_DATA.get();
 	}
 
 	public boolean isEnabled() {